refactor(frontend): move history view into separate folder
